pyproject.toml
[project]
name = "image-product-search"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.12",
    "httptools>=0.6.4",
    "numpy>=2.2.4",
    "open-clip-torch>=2.32.0",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.13.3",
    "streamlit>=1.44.1",
    "torch>=2.6.0",
    "transformers>=4.51.2",
    "uvicorn>=0.34.0",
    "uvloop>=0.21.0",
]
